portfolio/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
distlib==0.3.9
Django==5.1.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
filelock==3.16.1
packaging==24.2
pillow==11.0.0
pipenv==2024.4.0
platformdirs==4.3.6
python-decouple==3.8
python-http-client==3.3.7
sendgrid==6.11.0
setuptools==75.6.0
sqlparse==0.5.2
starkbank-ecdsa==2.2.0
tzdata==2024.2
virtualenv==20.28.0
whitenoise==6.8.2
